requirements.txt
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
attrs==23.1.0
autopep8==2.0.4
bcrypt==4.1.1
certifi==2023.11.17
click==8.1.7
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
fastapi-restful==0.5.0
gast==0.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.4
iniconfig==2.0.0
install==1.3.5
isort==5.12.0
mccabe==0.7.0
outcome==1.3.0.post0
packaging==23.2
passlib==1.7.4
platformdirs==4.1.0
pluggy==1.3.0
protobuf==4.21.12
psutil==5.9.6
pycodestyle==2.11.1
pydantic==1.10.13
pydantic_core==2.14.5
PyJWT==2.8.0
pylint==3.0.2
PyMySQL==1.1.0
pytest==7.4.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
